requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.112.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.2
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==2.0.1
openai==1.40.8
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
scikit-learn==1.5.1
scipy==1.13.1
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
sympy==1.13.2
threadpoolctl==3.5.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
xgboost==2.1.1
yarl==1.9.4
zipp==3.20.0

tensorflow~=2.16.2
werkzeug~=3.0.3